+++ sdk/appconfiguration/azure-appconfiguration/azure/appconfiguration/__init__.py
@@ -18,7 +18,13 @@ from ._models import (
     ConfigurationSnapshot,
     ConfigurationSettingsFilter,
 )
-from ._generated.models import SnapshotStatus
+from ._generated.models import (
+    SnapshotStatus,
+    ConfigurationSettingLabel,
+    LabelFields,
+    SnapshotFields,
+    ConfigurationSettingFields,
+)
 from ._version import VERSION
 from ._azure_appconfiguration_error import ResourceReadOnlyError
 
@@ -31,7 +37,11 @@ __all__ = [
     "SecretReferenceConfigurationSetting",
     "ConfigurationSnapshot",
     "SnapshotStatus",
+    "SnapshotFields",
+    "LabelFields",
+    "ConfigurationSettingFields",
     "ConfigurationSettingsFilter",
+    "ConfigurationSettingLabel",
     "FILTER_PERCENTAGE",
     "FILTER_TARGETING",
     "FILTER_TIME_WINDOW",