+++ sdk/appconfiguration/azure-appconfiguration/azure/appconfiguration/_generated/operations/_patch.py
@@ -2,6 +2,8 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 # ------------------------------------
+
+
 """Customize generated code here.
 
 Follow our quickstart for examples: https://aka.ms/azsdk/python/dpcodegen/python/customize
@@ -26,7 +28,6 @@ from ._azure_app_configuration_operations import (
     build_get_key_values_request,
 )
 from .. import models as _models
-from .._vendor import _convert_request
 
 
 class AzureAppConfigurationOperationsMixin(AzureAppConfigOpGenerated):
@@ -37,9 +38,10 @@ class AzureAppConfigurationOperationsMixin(AzureAppConfigOpGenerated):
         label: Optional[str] = None,
         after: Optional[str] = None,
         accept_datetime: Optional[str] = None,
-        select: Optional[List[Union[str, _models.KeyValueFields]]] = None,
+        select: Optional[List[Union[str, _models.ConfigurationSettingFields]]] = None,
         if_match: Optional[str] = None,
         if_none_match: Optional[str] = None,
+        tags: Optional[List[str]] = None,
         continuation_token: Optional[str] = None,
         **kwargs: Any
     ) -> Iterable["_models.KeyValue"]:
@@ -59,7 +61,7 @@ class AzureAppConfigurationOperationsMixin(AzureAppConfigOpGenerated):
         :type accept_datetime: str
         :param select: Used to select what fields are present in the returned resource(s). Default
          value is None.
-        :type select: list[str or ~azure.appconfiguration.models.KeyValueFields]
+        :type select: list[str or ~azure.appconfiguration.models.ConfigurationSettingFields]
         :param snapshot: A filter used get key-values for a snapshot. The value should be the name of
          the snapshot. Not valid when used with 'key' and 'label' filters. Default value is None.
         :type snapshot: str
@@ -69,6 +71,8 @@ class AzureAppConfigurationOperationsMixin(AzureAppConfigOpGenerated):
         :param if_none_match: Used to perform an operation only if the targeted resource's etag does
          not match the value provided. Default value is None.
         :type if_none_match: str
+        :param tags: A filter used to query by tags. Default value is None.
+        :type tags: list[str]
         :param str continuation_token: An opaque continuation token.
         :keyword callable cls: A custom type or function that will be passed the direct response
         :return: An iterator like instance of either KeyValue or the result of cls(response)
@@ -100,21 +104,21 @@ class AzureAppConfigurationOperationsMixin(AzureAppConfigOpGenerated):
                     select=select,
                     if_match=if_match,
                     if_none_match=if_none_match,
+                    tags=tags,
                     sync_token=self._config.sync_token,
                     api_version=api_version,
                     headers=_headers,
                     params=_params,
                 )
-                _request = _convert_request(_request)
                 path_format_arguments = {
                     "endpoint": self._serialize.url(
                         "self._config.endpoint", self._config.endpoint, "str", skip_quote=True
                     ),
                 }
                 _request.url = self._client.format_url(_request.url, **path_format_arguments)
-
             else:
                 # make call to next link with the client's api-version
+
                 _parsed_next_link = urllib.parse.urlparse(next_link)
                 _next_request_params = case_insensitive_dict(
                     {
@@ -126,7 +130,6 @@ class AzureAppConfigurationOperationsMixin(AzureAppConfigOpGenerated):
                 _request = HttpRequest(
                     "GET", urllib.parse.urljoin(next_link, _parsed_next_link.path), params=_next_request_params
                 )
-                _request = _convert_request(_request)
                 path_format_arguments = {
                     "endpoint": self._serialize.url(
                         "self._config.endpoint", self._config.endpoint, "str", skip_quote=True
@@ -148,13 +151,11 @@ class AzureAppConfigurationOperationsMixin(AzureAppConfigOpGenerated):
             map_error(status_code=response.status_code, response=response, error_map=error_map)
             error = self._deserialize.failsafe_deserialize(_models.Error, pipeline_response)
             raise HttpResponseError(response=response, model=error)
-
         response_headers = response.headers
         deserialized = self._deserialize("KeyValueListResult", pipeline_response)
 
         if cls:
             return cls(pipeline_response, deserialized, response_headers)
-
         return deserialized