+++ sdk/appconfiguration/azure-appconfiguration/samples/list_revision_sample.py
@@ -10,7 +10,7 @@
 FILE: list_revision_sample.py
 
 DESCRIPTION:
-    This sample demos list revision operations for app configuration
+    This sample demos how to get configuration setting revision history synchronously.
 
 USAGE: python list_revision_sample.py
 
@@ -19,13 +19,12 @@ USAGE: python list_revision_sample.py
 """
 import os
 from azure.appconfiguration import AzureAppConfigurationClient, ConfigurationSetting
-from util import print_configuration_setting
 
 
 def main():
     CONNECTION_STRING = os.environ["APPCONFIGURATION_CONNECTION_STRING"]
 
-    # Create app config client
+    # Create an app config client
     client = AzureAppConfigurationClient.from_connection_string(CONNECTION_STRING)
 
     config_setting = ConfigurationSetting(
@@ -37,14 +36,13 @@ def main():
     returned_config_setting.content_type = "new content type"
     client.set_configuration_setting(config_setting)
 
-    items = client.list_revisions(key_filter="MyKey")
+    # [START list_revisions]
+    items = client.list_revisions(key_filter="MyKey", tags_filter=["my tag=my tag value"])
     for item in items:
-        print_configuration_setting(item)
-        print("")
+        print(item)
+    # [END list_revisions]
 
-    client.delete_configuration_setting(
-        key="MyKey",
-    )
+    client.delete_configuration_setting(key="MyKey")
 
 
 if __name__ == "__main__":