+++ sdk/appconfiguration/azure-appconfiguration/samples/list_revision_sample_async.py
@@ -7,12 +7,12 @@
 # --------------------------------------------------------------------------
 
 """
-FILE: list_revision_async_sample.py
+FILE: list_revision_sample_async.py
 
 DESCRIPTION:
-    This sample demos list revision operations for app configuration
+    This sample demos how to get configuration setting revision history asynchronously.
 
-USAGE: python list_revision_async_sample.py
+USAGE: python list_revision_sample_async.py
 
     Set the environment variables with your own values before running the sample:
     1) APPCONFIGURATION_CONNECTION_STRING: Connection String used to access the Azure App Configuration.
@@ -21,13 +21,12 @@ import asyncio
 import os
 from azure.appconfiguration import ConfigurationSetting
 from azure.appconfiguration.aio import AzureAppConfigurationClient
-from util import print_configuration_setting
 
 
 async def main():
     CONNECTION_STRING = os.environ["APPCONFIGURATION_CONNECTION_STRING"]
 
-    # Create app config client
+    # Create an app config client
     client = AzureAppConfigurationClient.from_connection_string(CONNECTION_STRING)
 
     config_setting = ConfigurationSetting(
@@ -39,14 +38,11 @@ async def main():
     returned_config_setting.content_type = "new content type"
     await client.set_configuration_setting(config_setting)
 
-    items = client.list_revisions(key_filter="MyKey")
+    items = client.list_revisions(key_filter="MyKey", tags_filter=["my tag=my tag value"])
     async for item in items:
-        print_configuration_setting(item)
-        print("")
+        print(item)
 
-    await client.delete_configuration_setting(
-        key="MyKey",
-    )
+    await client.delete_configuration_setting(key="MyKey")
 
 
 if __name__ == "__main__":